commit 3b
Removed HTML structure and comments from the CSS file. (my code editor wouldn't recognize it as valid code without removing it, so I think it just applies to the HTML part and not CSS). completed 'box styles' section of the challenge, referencing the 'body' and 'ul' elements of the HTML